refactor(page): Move 'page' struct to 'manga' package
