+++ internal/manga/page.go
@@ -1,4 +1,4 @@
-package downloader
+package manga
 
 import (
 	"fmt"
@@ -12,13 +12,13 @@ import (
 type Page struct {
 	Url, Name string
 	ext       string // file extension from MIME type
-	bytes     []byte
+	Bytes     []byte
 }
 
-func (p *Page) download() (*Page, error) {
+func (p *Page) Download(delayStr string, pageRetries int) (*Page, error) {
 	logging.Debugln("Starting download of page: ", p.Filename)
 
-	dur, err := time.ParseDuration(config.Delay.Page)
+	dur, err := time.ParseDuration(delayStr)
 	if err != nil {
 		return p, err
 	}
@@ -26,11 +26,11 @@ func (p *Page) download() (*Page, error) {
 
 	imageBytesResp, _ := rester.New().GetBytes(p.Url,
 		map[string]string{},
-		[]rester.QueryParam{}).Do(config.PageRetries, "100ms")
-	p.bytes = imageBytesResp.([]byte)
+		[]rester.QueryParam{}).Do(pageRetries, "100ms")
+	p.Bytes = imageBytesResp.([]byte)
 
 	// Get mime type
-	mimeType := http.DetectContentType(p.bytes)
+	mimeType := http.DetectContentType(p.Bytes)
 	ext, err := mime.ExtensionsByType(mimeType)
 	if err != nil {
 		return p, err
@@ -38,7 +38,7 @@ func (p *Page) download() (*Page, error) {
 
 	p.ext = ext[len(ext)-1] // Hacky way to get image/jpeg to be .jpg, but keep everything else the same
 
-	logging.Debugln("Downloaded page. Byte length: ", len(p.bytes))
+	logging.Debugln("Downloaded page. Byte length: ", len(p.Bytes))
 
 	return p, nil
 }
